+++ swarm_copy/tools/__init__.py
@@ -1,7 +1,11 @@
 """Tools package."""
 
+from swarm_copy.tools.bluenaas_memodel_getall import MEModelGetAllTool
+from swarm_copy.tools.bluenaas_memodel_getone import MEModelGetOneTool
+from swarm_copy.tools.bluenaas_scs_getall import SCSGetAllTool
+from swarm_copy.tools.bluenaas_scs_getone import SCSGetOneTool
+from swarm_copy.tools.bluenaas_scs_post import SCSPostTool
 from swarm_copy.tools.electrophys_tool import ElectrophysFeatureTool, FeatureOutput
-from swarm_copy.tools.get_me_model_tool import GetMEModelTool
 from swarm_copy.tools.get_morpho_tool import GetMorphoTool, KnowledgeGraphOutput
 from swarm_copy.tools.kg_morpho_features_tool import (
     KGMorphoFeatureOutput,
@@ -22,6 +26,9 @@ from swarm_copy.tools.resolve_entities_tool import (
 from swarm_copy.tools.traces_tool import GetTracesTool, TracesOutput
 
 __all__ = [
+    "SCSGetAllTool",
+    "SCSGetOneTool",
+    "SCSPostTool",
     "BRResolveOutput",
     "ElectrophysFeatureTool",
     "FeatureOutput",
@@ -31,10 +38,11 @@ __all__ = [
     "KGMorphoFeatureTool",
     "KnowledgeGraphOutput",
     "LiteratureSearchTool",
+    "MEModelGetAllTool",
+    "MEModelGetOneTool",
     "MorphologyFeatureOutput",
     "MorphologyFeatureTool",
     "ParagraphMetadata",
     "ResolveEntitiesTool",
     "TracesOutput",
-    "GetMEModelTool",
 ]